requirements.txt
flask
newsapi-python
flask-restful
praw
python-decouple